+++ examples/function_call.rs
@@ -18,6 +18,7 @@ fn main() {
         let v1 = builder.ins().const_int(1, 0);
         let v2 = builder.ins().const_int(42, 9001);
         let v3 = builder.ins().call(ext, vec![v1, v2]);
+        let v3 = builder.dfg().inst_result(v3);
         builder.ins().ret_value(v3);
     }
     println!("{}", func.dump());